Brings in the latest mono/skia 'skiasharp' branch updates on top of the
m147 work-in-progress:

  * Add variable font C API for typeface variation queries and cloning (#185)
  * Add sk_fontmgr_legacy_create_typeface C API (#198, #195) — replaces
    the old sk_default_fontmgr.{cpp,h} with a single C entry point that
    forwards to SkFontMgr::legacyMakeTypeface(). Fixes the Android
    default-typeface SIGSEGV.
  * Use int32_t instead of long for SkManagedStream move offset
    (mono/skia#186) — fail-loud on overflow instead of silent truncation
    of platform-dependent C long.
  * Update libpng to 1.6.58
  * Update libexpat to 2.7.5

Conflict resolutions:
  - DEPS: keep m147's freetype roll, take origin/skiasharp's libexpat 2.7.5
    (and libpng 1.6.58 auto-merged the same way).
  - src/c/sk_default_fontmgr.cpp: accept origin/skiasharp's deletion;
    their legacyMakeTypeface refactor obviates our m147 carry of that file.
  - src/c/sk_typeface.cpp: re-apply our m139 SkFontScanner fix to the
    Android and FontConfig SkFontMgr_New_* call sites, which moved into
    this file as part of their refactor.

+++ include/xamarin/SkManagedStream.h
@@ -47,7 +47,7 @@ public:
     typedef bool              (*RewindProc)      (      SkManagedStream* s, void* context);
     typedef size_t            (*GetPositionProc) (const SkManagedStream* s, void* context);
     typedef bool              (*SeekProc)        (      SkManagedStream* s, void* context, size_t position);
-    typedef bool              (*MoveProc)        (      SkManagedStream* s, void* context, long offset);
+    typedef bool              (*MoveProc)        (      SkManagedStream* s, void* context, int32_t offset);
     typedef size_t            (*GetLengthProc)   (const SkManagedStream* s, void* context);
     typedef SkManagedStream*  (*DuplicateProc)   (const SkManagedStream* s, void* context);
     typedef SkManagedStream*  (*ForkProc)        (const SkManagedStream* s, void* context);
